improve user access manager extensibility
